+++ R/Support.R
@@ -8,10 +8,9 @@
 #' 
 #' 
 #' @param reference A tree of class phylo, or a character vector specifying its splits 
-#'                     (as obtained through [TreeSplits])
-#' @param forest a list of trees of class phylo, or a multiPhylo object; or a table
-#'               enumerating the occurrences of each split in that forest 
-#'                     (as obtained through [ForestSplits])
+#'                     (as obtained through [Quartet::Tree2Splits])
+#' @param forest a list of trees of class phylo, or a multiPhylo object; or a
+#' list of their constituent splits (as obtained through [Quartet::Tree2Splits])
 #'                     
 #' @return Number of trees in `forest` that contain each split in `reference`. 
 #'         if `reference` is a tree of class phylo, then the sequence will
@@ -19,9 +18,6 @@
 #'         Note that the three nodes at the root of the tree correspond to a 
 #'         single split; see the example for how these might be plotted on a tree.
 #' 
-#' @author Martin R. Smith
-#' @importFrom phangorn Descendants
-#' @export
 #' 
 #' @examples {
 #'   library(ape) # for functions rtree & nodelabels
@@ -32,28 +28,54 @@
 #'   forest <- list(tree1, tree2, tree2, tree3, rtree(7))
 #'   
 #'   # Simple, but means counting each split in the forest twice:
-#'   SplitFrequency(tree1, forest)
+#'   tree1Freqs <- SplitFrequency(tree1, forest)
 #'   SplitFrequency(tree2, forest)
 #'   
-#'   # Takes more typing, but caches the forest splits to save time at runtime:
-#'   forestSplits <- ForestSplits(forest)
-#'   tree1Freqs <- SplitFrequency(tree1, forestSplits)
-#'   tree2Freqs <- SplitFrequency(tree2, forestSplits)
-#'   
 #'   plot(tree1)
-#'   nodelabels(c('-', '-', # The first two nodes do not denote unique splits
-#'                tree1Freqs))
+#'   nodelabels(tree1Freqs, node=as.integer(names(tree1Freqs)))
 #' }
+#' 
+#' @author Martin R. Smith
+#' @export
 SplitFrequency <- function(reference, forest) {
   if (class(reference) %in% c('list', 'phylo')) {
-    reference <- TreeSplits(reference)
+    referenceSplits <- Tree2Splits(reference)
+  } else {
+    referenceSplits <- reference
   }
-  if (class(forest) %in% c('list', 'phylo', 'multiPhylo')) {
-    forest <- ForestSplits(forest)
+  if (class(forest) == 'phylo') {
+    forestSplits <- Tree2Splits(forest) 
+  } else if (class(forest) %in% c('list', 'multiPhylo')) {
+    forestSplits <- lapply(forest, Tree2Splits)
+  } else {
+    forestSplits <- forest
   }
   
+  ret <- rowSums(vapply(forestSplits, function (cf) SplitsRepeated(referenceSplits, cf),
+         logical(ncol(referenceSplits))))
+  names(ret) <- colnames(referenceSplits)
   # Return:
-  forest[reference]
+  ret
+}
+
+#' Are Splits Repeated?
+#' 
+#' Determines whether any of the splits described in `original` are present in
+#' `matches`.
+#' 
+#' @param original,matches Logical matrix describing the bipartition splits
+#' defined by a tree (see [Quartet:Tree2Splits] for format).  The first row
+#' of `original` must be `FALSE` for all splits.
+#' 
+#' @keywords internal
+#' @export
+#' @author Martin R. Smith
+#' @return `SplitsRepeated` returns a logical vector of length `ncol(original)`,
+#' specifying whether each split listed in `original` also appears in `matches`.
+SplitsRepeated <- function (original, matches) {
+  comparison <- matches[rownames(original), , drop=FALSE]
+  comparison[, comparison[1, ]] <- !comparison[, comparison[1, ]]
+  duplicated(t(cbind(comparison, original)))[-seq_len(ncol(matches))]
 }
 
 #' @describeIn SplitFrequency Assign a unique integer to each split
@@ -64,18 +86,19 @@ SplitFrequency <- function(reference, forest) {
 #'  of 2 to be associated with each tip in turn
 #' @export
 SplitNumber <- function (tips, tree, tipIndex, powersOf2) {
+  .Deprecated("SplitFrequency")
   included <- tipIndex %in% tree$tip.label[tips]
   as.character(min(c(sum(powersOf2[included]), sum(powersOf2[!included]))))
 }
 
 #' @describeIn SplitFrequency Frequency of splits in a given forest of trees
-#' @importFrom gmp as.bigz
 #' @export
-ForestSplits <- function (forest) {
-  if (class(forest) == 'phylo') forest <- list(forest)
+ForestSplits <- function (forest, powersOf2) {
+  .Deprecated("SplitFrequency")
+  if (class(forest) == 'phylo') forest <- structure(list(forest), class='multiPhylo')
   tipIndex <- sort(forest[[1]]$tip.label)
   nTip <- length(tipIndex)
-  powersOf2 <- as.bigz(2L ^ (seq_len(nTip) - 1L))
+  
   # Return:
   table(vapply(forest, function (tr) {
     # +2: Don't consider root node (not a node) or first node (duplicated)
@@ -88,6 +111,7 @@ ForestSplits <- function (forest) {
 #' @importFrom gmp as.bigz
 #' @export
 TreeSplits <- function (tree) {
+  .Deprecated("Quartet::Tree2Splits")
   tipIndex <- sort(tree$tip.label)
   nTip <- length(tipIndex)
   powersOf2 <- as.bigz(2L ^ (seq_len(nTip) - 1L))