+++ src/TreeSearch-init.c
@@ -2,9 +2,8 @@
 
 #include <Rmath.h>
 #include <math.h>
-#include <R.h> 
 #include <Rinternals.h>
-#include <stdlib.h> // for NULL
+#include <stdlib.h> /* for NULL */
 #include <R_ext/Rdynload.h>
 
 #include "ape_reorder.h"
@@ -15,6 +14,9 @@
 #include "RMorphy.h"
 #include "build_postorder.h"
 
+
+extern SEXP _TreeSearch_phangorn_bipCPP(SEXP, SEXP);
+
 static const R_CMethodDef cMethods[] = {
   {"order_edges_number_nodes", (DL_FUNC) &order_edges_number_nodes, 3, order_edges_number_nodes_t},
   {"ape_neworder_phylo",       (DL_FUNC) &ape_neworder_phylo, 6, ape_neworder_phylo_t},
@@ -24,6 +26,7 @@ static const R_CMethodDef cMethods[] = {
 };
 
 static const R_CallMethodDef callMethods[] = {
+  {"_TreeSearch_phangorn_bipCPP",   (DL_FUNC) &_TreeSearch_phangorn_bipCPP, 2},
   {"_R_wrap_mpl_new_Morphy",        (DL_FUNC) &_R_wrap_mpl_new_Morphy, 0},
   {"_R_wrap_mpl_delete_Morphy",     (DL_FUNC) &_R_wrap_mpl_delete_Morphy, 1},
   {"_R_wrap_mpl_init_Morphy",       (DL_FUNC) &_R_wrap_mpl_init_Morphy, 3},