+++ src/build_postorder.h
@@ -2,8 +2,9 @@
 #include <stdio.h>
 #include "RMorphy.h"
 
-// Random number generator from http://www.cse.yorku.ca/~oz/marsaglia-rng.html
-// 1+random_int%10 generates an integer from 1 to 10 [MWC renamed to random_int]
+/* Random number generator from http://www.cse.yorku.ca/~oz/marsaglia-rng.html
+/ 1+random_int%10 generates an integer from 1 to 10 [MWC renamed to random_int]
+*/
 #define znew (z=36969*(z&65535)+(z>>16))
 #define wnew (w=18000*(w&65535)+(w>>16))
 #define random_int ((znew<<16)+wnew)
@@ -20,7 +21,7 @@ void insert_tip_below (const int *new_tip,
   if (left[old_parent] == *add_below) {
     left[old_parent] = *new_node;
   } else {
-    // The same, but on the right
+    /* The same, but on the right */
     right[old_parent] = *new_node;
   }
   parent_of[*new_node] = old_parent;
@@ -32,19 +33,20 @@ void insert_tip_below (const int *new_tip,
   parent_of[*add_below] = *new_node;
 }
 
-// parent_of, left and right have been initialized with a two-taxon tree with tips 0 & 1
-// left and right point n_tip _before_ left and right, so we don't need to subtract n_tip each time
-// We arbitrarily choose to root our tree on tip 0, so never add to that edge or the 
-// "dummy" root edge.
+/* parent_of, left and right have been initialized with a two-taxon tree with tips 0 & 1
+ *left and right point n_tip _before_ left and right, so we don't need to subtract n_tip each time
+ *We arbitrarily choose to root our tree on tip 0, so never add to that edge or the 
+ *"dummy" root edge.
+ */
 void build_tree(int *parent_of, int *left, int *right, const int *n_tip) {
   int tip_to_add, add_below, new_node;
   for (tip_to_add = 3; tip_to_add < *n_tip; tip_to_add++) {
     new_node = tip_to_add + *n_tip - 1;
-    add_below = 1 + random_int % (tip_to_add + tip_to_add - 3); // +1 to avoid edge 0
-    if (add_below < tip_to_add) { // Adding below a tip
+    add_below = 1 + random_int % (tip_to_add + tip_to_add - 3); /* +1 to avoid edge 0 */
+    if (add_below < tip_to_add) { /* Adding below a tip */
       insert_tip_below(&tip_to_add, &add_below, &new_node, parent_of, left, right);
-    } else { // Adding below an existing node
-      add_below += *n_tip - tip_to_add + 1; // +1 so we never touch dummy root edge
+    } else { /* Adding below an existing node */
+      add_below += *n_tip - tip_to_add + 1; /* +1 so we never touch dummy root edge */
       insert_tip_below(&tip_to_add, &add_below, &new_node, parent_of, left, right);
     }
   }
@@ -60,7 +62,7 @@ void move_to_node(const int *old_node_id, int *new_parent, int *new_left, int *n
     move_to_node(&old_right[*old_node_id], new_parent, new_left, new_right,
                                  old_parent, old_left, old_right, 
                                  next_label, n_tip);    
-  } else if (new_node_id != *old_node_id) { // Otherwise no change
+  } else if (new_node_id != *old_node_id) { /* Otherwise no change */
     new_parent[old_right[*old_node_id]] = new_node_id;
     new_right[new_node_id] = old_right[*old_node_id];
   }
@@ -70,7 +72,7 @@ void move_to_node(const int *old_node_id, int *new_parent, int *new_left, int *n
     move_to_node(&old_left[*old_node_id], new_parent, new_left, new_right,
                                  old_parent, old_left, old_right, 
                                  next_label, n_tip);    
-  } else if (new_node_id != *old_node_id) { // Otherwise no change
+  } else if (new_node_id != *old_node_id) { /* Otherwise no change */
     new_parent[old_left[*old_node_id]] = new_node_id;
     new_left[new_node_id] = old_left[*old_node_id];
   }
@@ -103,18 +105,18 @@ void renumber_postorder(int *parent_of, int *left, int *right, const int *n_tip)
 
 void random_tree(int *parent_of, int *left, int *right, const int *n_tip) {
   if (*n_tip < 3) {
-        // Initialize with 2-tip tree
+        /* Initialize with 2-tip tree */
        parent_of[0] = *n_tip;
        parent_of[1] = *n_tip;
-  parent_of[*n_tip] = *n_tip; // Root is its own parent
+  parent_of[*n_tip] = *n_tip; /* Root is its own parent */
             left[0] = 0;
            right[0] = 1;
   } else {
-    // Initialize with 3-tip tree, arbitrarily rooted on tip 0
+    /* Initialize with 3-tip tree, arbitrarily rooted on tip 0 */
              parent_of[0] = *n_tip;
              parent_of[1] = *n_tip + 1;
              parent_of[2] = *n_tip + 1;
-        parent_of[*n_tip] = *n_tip; // Root is its own parent
+        parent_of[*n_tip] = *n_tip; /* Root is its own parent */
     parent_of[*n_tip + 1] = *n_tip;
                   left[0] = 0;
                   left[1] = 1;
@@ -152,7 +154,11 @@ extern SEXP RANDOM_TREE_SCORE(SEXP ntip, SEXP MorphyHandl) {
   const int n_tip = INTEGER(ntip)[0];
   Morphy handl = R_ExternalPtrAddr(MorphyHandl);
   SEXP RESULT = PROTECT(allocVector(INTSXP, 1));
-  int *score;
+  int *score,
+    *parent_of = calloc(n_tip + n_tip - 1 , sizeof(int)),
+    *left = calloc(n_tip - 1         , sizeof(int)),
+    *right = calloc(n_tip - 1         , sizeof(int));
+    
   score = INTEGER(RESULT);
   *score = 0;
   if (n_tip < 2) {
@@ -161,11 +167,6 @@ extern SEXP RANDOM_TREE_SCORE(SEXP ntip, SEXP MorphyHandl) {
     return(RESULT);
   }
   
-  // NOTE: malloc here causes segfault.
-  //       Does this mean that we're accessing uninitialzied values somewhere?
-  int *parent_of = calloc(n_tip + n_tip - 1 , sizeof(int)),
-           *left = calloc(n_tip - 1         , sizeof(int)),
-          *right = calloc(n_tip - 1         , sizeof(int));
   
   random_tree(parent_of, left, right, &n_tip);
   morphy_length(parent_of, left, right, handl, score);