+++ src/renumber_tree.h
@@ -10,15 +10,18 @@ extern void order_edges_number_nodes(int *parent, int *child, const int *n_edge)
   int i, q_pos = 0, o_node, next_node;
   const int n_node = *n_edge / 2;
   const int n_allnodes = *n_edge + 1L, root_node = n_node + 2L;
-  int * start_p = calloc(*n_edge, sizeof(int)); // calloc zero-initializes
-  int * start_c = calloc(*n_edge, sizeof(int));
-  int * child_l = calloc( n_node, sizeof(int));
-  int * child_r = calloc( n_node, sizeof(int));
-  int * queue_p = calloc( n_node, sizeof(int));
-  int * queue_c = calloc( n_node, sizeof(int));
-  // TODO check that calloc has returned a non-null pointer; clean-up and exit if calloc has failed
+  int * start_p = calloc(*n_edge, sizeof(int)), /* calloc zero-initializes */
+      * start_c = calloc(*n_edge, sizeof(int)),
+      * child_l = calloc( n_node, sizeof(int)),
+      * child_r = calloc( n_node, sizeof(int)),
+      * queue_p = calloc( n_node, sizeof(int)),
+      * queue_c = calloc( n_node, sizeof(int)),
+      * renumber = calloc(n_allnodes, sizeof(int));
+  /* TODO check that calloc has returned a non-null pointer; 
+   clean-up and exit if calloc has failed */
+    
   for (i = 0; i < *n_edge; i++) {
-    // Initialize
+    /* Initialize */
     start_p[i] = parent[i];
     start_c[i] = child[i];
     q_pos = parent[i] - root_node;
@@ -36,11 +39,11 @@ extern void order_edges_number_nodes(int *parent, int *child, const int *n_edge)
   o_node = root_node;
   q_pos = 0;
   for (i = 0; i < *n_edge; i++) {
-    if (o_node < root_node) { // We've just reached a tip
+    if (o_node < root_node) { /* We've just reached a tip */
       parent[i] = queue_p[--q_pos];
       child[i] = queue_c[q_pos];
       o_node = child[i];
-    } else { // We're at an internal node
+    } else { /* We're at an internal node */
       parent[i] = o_node;
       child[i]  = child_l[o_node - root_node];
       queue_p[q_pos] = o_node;
@@ -55,8 +58,7 @@ extern void order_edges_number_nodes(int *parent, int *child, const int *n_edge)
   free(queue_p);
   free(queue_c);
   
-  // Now number nodes:
-  int * renumber = calloc(n_allnodes, sizeof(int));
+  /* Now number nodes: */
   if (renumber != NULL) {
     next_node = root_node;
     for (i = 0; i < n_allnodes; i++) renumber[i] = i + 1;